Add webpack karma config as reference
This adds a configuration for Webpack as a reference to anyone using it
over Browserify since it requires some special configuration to work.
There is a TODO task here to make the tests run with both Webpack and
Browserify configs in series, but that will take some work as the
current test setup won't allow that easily. For now, having this example
will help others to figure out how to get request working with Webpack.

This closes #1529